Improve seeds protection for review apps.

+++ db/seeds.rb
@@ -6,9 +6,8 @@
 #   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
 #   Mayor.create(name: 'Emanuel', city: cities.first)
 
-fail "Seeds can be run only in development" unless Rails.env.development?
 
-Client.destroy_all
+Client.destroy_all if Rails.env.development?
 
 admin = Client.create( name: "admin", password: "password", email: "admin@gorgu.net", active: true)
 admin.add_role :gram_admin
